- added const max allowed delta (tolerance) for the brightness search - switched from int-mean to float mean to decrease rounding errors - added comments / better readability

+++ include/pylon_camera/pylon_camera_node.h
@@ -144,7 +144,7 @@ protected:
      * Calculates the mean brightness of the image
      * @return the mean brightness of the image
      */
-    virtual int calcCurrentBrightness();
+    virtual float calcCurrentBrightness();
 
     /**
      * Getter for the current exposure
